Fixes to Windows recipe instructions.
Pins the Visual Studio version to 2017 and invokes vcvarsall.bat each
time so as to ensure that CMake will find the compiler when multiple
versions of Visual Studio are installed (i.e. 2017 and 2019 both on
build host0.
Additional fixes to the Windows clamav recipe instructions to fix the
copy and fix setting/restoring of environment variables for the build.