libxml2: update to 2.9.12, remove patch directory
libxml2 finally has a release that patches the CVEs we used to need
patches for, plus 2 new CVEs that we hadn't yet added patches for.

Note: this update uses a download from their gitlab for the release
materials because the autotools-generated release tarball is missing a
file that is required on Windows.