libxml2: update to 2.9.12, remove patch directory
libxml2 finally has a release that patches the CVEs we used to need
patches for, plus 2 new CVEs that we hadn't yet added patches for.

On Windows this recipe now uses cmake to install instead of copying the
files manually. This is required because the libxml2 CMake build system
appears to only create the xmlversion.h file on Windows in the install.
Of course, this is the best way to install the files anyways, so we
should do the same for all other Windows CMake recipes.

Note: this update uses a download from their gitlab for the release
materials because the autotools-generated release tarball is missing a
file that is required on Windows.