openssl: fix copy recursion issue in darwin recipe
There is a copy recursion issue with the arm64 directory.  Instead,
we'll copy up a directory I guess so it goes in the base work directory
instead of as a subdirectory.

I found this issue while trying to figure out why we're seeing linker
issues with the openssl 3.1 static libs on darwin.  This doesn't solve
it, but may as well be fixed.