ncurses: also build libtinfo
Building with ncurses is adding a dependency on libtinfo although we're
not building/installing libtinfo!!

When doing a clamav build with static dependencies, that ends up
depending on a SHARED libtinfo from the host instead of building in our
own static libtinfo librar!!!

That looks like this in the clamav CMake configuration summary:

-- Application Extra Dependencies --
        GUI support:
            ncurses         /usr/include
                            /home/micah/.mussels/install/host-static/lib/libncurses.a;/usr/lib/x86_64-linux-gnu/libtinfo.so

That means that clamdtop will not work without libtinfo installed
(somehow), or may not work at all if the expected path does not match
the way it is installed on the host.

This commit enables building/installing libtinfo.a or .so for ncurses
builds.