libcurl, Windows: disable libidn2 disable pkg-config
Curl has an option to use libidn2 in version 8.9.0.

Pkg-config may unexpectedly find it and use it.

I encountered this on Windows because libidn2 was installed by
Strawberry Perl, under C:\Strawberry. Pkg-config found that and tried to
build it into Curl, which is problematic in part because we don't want
to include or ship libidn2 with our software.

But also problematic because libidn2 provided by Strawberry Perl used a
different OpenSSL incompatible with the one we're compiling against.

This commit explicitly disables pkg-config and libidn2 for libcurl
builds.
TODO: Perhaps disable pkg-config for all recipes and all platforms?
Mussels builds should only draw from other mussels include/lib files.