libcurl: Disable libidn2 for all builds
Automatic library detection finds brew-installed libs.
This is primarily to fix builds on GitHub Actions where
the system provides all sorts of stuff we didn't request.

On macOS, specifically, this causes failures because we build
universal binaries, and brew binaries are not universal and
lack symbols for othe other platform.