Update curl, nghttp2, openssl, pcre2, libxml2
libcurl from 8.15 to 8.16

libnghttp2 from 1.66 to 1.67

openssl from 3.2.5 to 3.2.6

libpcre2 from 10.45 to 10.46

libxml2 from 2.14.5 to 2.15.0

Also update and extend the clamav recipes, removing patch versions to
make them more generic and adding 1.4, 1.5, and 1.6 (for main/dev).

Also change the URL for the ncurses 6.5 tarball as the FTP server for
the old server appears to be offline.